requirements.txt
aiolimiter==1.1.0
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
astor==0.8.1
attrdict==2.0.1
Babel==2.14.0
bce-python-sdk==0.8.98
binaryornot==0.4.4
blessed==1.20.0
blinker==1.7.0
bs4==0.0.1
crcmod==1.7
cssselect==1.2.0
cssutils==2.9.0
Cython==3.0.8
dashscope==1.14.0
et-xmlfile==1.1.0
Flask==3.0.0
flask-babel==4.0.0
future==0.18.3
hpack==4.0.0
httpx-sse==0.4.0
imageio==2.33.1
imgaug==0.4.0
itsdangerous==2.1.2
jieba==0.42.1
jmespath==0.10.0
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
langchainhub==0.1.14
langserve==0.0.39
langsmith==0.0.80
lazy_loader==0.3
lmdb==1.4.1
munkres==1.1.4
nvidia-ml-py==12.535.77
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
oss2==2.18.4
paddleocr==2.6.0.1
paddlepaddle==2.6.0
peft==0.4.0
pip-search==0.0.12
premailer==3.10.0
protobuf==4.24.4
pyarrow==14.0.2
pyclipper==1.3.0.post5
pycryptodome==3.19.1
pydantic==1.10.13
pydantic-settings==2.1.0
PyMuPDF==1.21.1
python-docx==1.1.0
python-dotenv==1.0.0
python-pptx==0.6.23
qianfan==0.2.5
rapidfuzz==3.6.1
rarfile==4.1
redis==5.0.1
rjieba==0.1.11
scikit-image==0.22.0
shapely==2.0.2
sse-starlette==1.8.2
text-unidecode==1.3
tifffile==2023.12.9
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
types-requests==2.31.0.20240106
visualdl==2.5.3
Werkzeug==3.0.1
XlsxWriter==3.1.9
